Don't remove the sprite file if it's going to be replaced.

+++ lib/compass/sass_extensions/sprites/sprite_methods.rb
@@ -69,6 +69,7 @@ module Compass
         
         def cleanup_old_sprites
           Dir[File.join(Compass.configuration.generated_images_path, "#{path}-s*.png")].each do |file|
+            next if file[filename]
             log :remove, file
             FileUtils.rm file
             Compass.configuration.run_sprite_removed(file)
